agents/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dataclasses-json==0.5.7
Flask==2.3.1
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-search-results==2.4.1
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.74
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.24.1
openai==0.26.4
packaging==23.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
pyparsing==3.0.9
python-engineio==4.4.1
python-socketio==5.8.0
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.46
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.3.2
yarl==1.8.2
